Exercises 4.15-4.22

+++ part4/blog/controllers/blogs.js
@@ -1,13 +1,29 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+
+
+
 
 blogsRouter.get('/',  async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1})
+
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
 
   if (!body.title || !body.url) {
     return response.status(400).json({error: 'title or url missing'})
@@ -17,16 +33,33 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: (body.hasOwnProperty('likes')) ? body.likes : 0
+    likes: (body.hasOwnProperty('likes')) ? body.likes : 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
-  return response.json(savedBlog)
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  return response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+
+  const blogToRemove = await Blog.findByIdAndRemove(request.params.id)
+  if (blogToRemove.user.toString() === user.id.toString()) {
+    return response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'Authorization to delete blog missing' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
